+++ plugins/wt-deep-search/src/index.js
@@ -1,0 +1,5 @@
+export { detectProviders } from './detect.js';
+export { route } from './route.js';
+export { searchBrave } from './clients/brave.js';
+export { searchExa } from './clients/exa.js';
+export { ProviderFailure, createExhaustionMemory } from './provider-failure.js';